finished restructuring of DirectoryRenameFiles to minimize globals
