+++ DirectoryRenameFiles_v2.py
@@ -8,17 +8,28 @@ from tkinter import filedialog
 import os
 
 
-# the frame class?
-class Example(Frame):
+class FileRename(tk.Frame):
 
-    # 'global' variables to hold selected state of buttons
+    # variables to hold selected state of buttons
     bIncludeSubdirs = False
     bPrefix = False
     bSuffix = False
     bInsertB = False
     bInsertA = False
     bReplace = False
-
+    prefix = ""
+    suffix = ""
+    insertB = ""
+    insertA = ""
+    replace = ""
+    findtext = ""
+    # initialize as list types
+    subDirList = []
+    oldFileList = []
+    newFileList = []
+    oldFilePathList = []
+    newFilePathList = []
+    tempFilePathList = []
 
     # some shit that we do because it's the 'right' way to do things
     def __init__(self):
@@ -26,23 +37,57 @@ class Example(Frame):
 
         self.initUI()
 
-
+    # the main frame
     def initUI(self):
 
+        # declare all functions first
+        # function for browsing to a directory
+        def BrowseFolder():
+            global userDirectory
+            global sUserDirectory
+            # prompt user
+            userDirectory=filedialog.askdirectory()
+            sUserDirectory.set(userDirectory)
+            # check if user cancelled prompt - this doesn't actually work - not sure what cancelled dialog returns
+            if userDirectory is "":
+                userDirectory = "nothing selected"
+                sUserDirectory.set("")
+            # debug print
+            #print("userDirectory: " + self.userDirectory)
+            getfilelist(userDirectory)
 
-        def saveFile(i):
-            global oldfilepathlist
-            global newfilepathlist
-            os.rename(tempfilepathlist[i], newfilepathlist[i])
-
-
-        # function to check if a string variable is not empty
-        def isNotBlank (myString):
-            if myString and myString.strip():
-                #myString is not None AND myString is not empty or blank
-                return True
-            #myString is None OR myString is empty or blank
-            return False
+        # function for getting all files within a directories and its subs
+        def getfilelist(directory):
+            global sOldFileList
+            global soldFilePathList
+            self.subDirList.clear()
+            self.oldFileList.clear()
+            self.oldFilePathList.clear()
+            self.newFileList.clear()
+            self.newFilePathList.clear()
+            # clear new file name display
+            snewFileList.set("")
+            for subdir, dirs, files in os.walk(directory):
+                for file in files:
+                    filepath = subdir + os.sep + file
+                    # add subdir + ending slash to global subdir variable
+                    self.subDirList.append(subdir + os.sep)
+                    # add file to global list variable
+                    self.oldFileList.append(file)
+                    # append null string to keep list the same size
+                    self.newFileList.append("")
+                    self.oldFilePathList.append(filepath)
+                    self.newFilePathList.append("")
+                    self.tempFilePathList.append("")
+                if not self.bIncludeSubdirs:
+                    # if user selected not to include subdirectories,
+                    #  break out of loop after going through files
+                    break
+            # get string list of files to display to the user
+            soldFileList.set("\n".join(str(x) for x in self.oldFileList))
+            # get string list of filepaths to display to the user
+            soldFilePathList.set("\n".join(str(x) for x in self.oldFilePathList))
+            updatebody(scrollable_body)
 
 
         # function to rename a list of files
@@ -50,135 +95,89 @@ class Example(Frame):
         # will need to access global file name/path lists
         def RenameStuff(sPrefix, sSuffix, sInsertB, sInsertA, sReplace, texttofind):
             # get directory again in case user is running renameStuff again without re-browsing to folders
-            global userdirectory
-            getfilelist(userdirectory)
-            global tempfilepathlist
-            global oldfilelist
-            global oldfilepathlist
-            global subdirlist
-            global resavefile
+            global userDirectory
+            global snewFileList
+
+            getfilelist(userDirectory)
             tempfilename = ""
 
             # if user selected to insert or replace text, need to find the provided string
             bFind = False
-            if Example.bInsertB or Example.bInsertA or Example.bReplace:
+            if self.bInsertB or self.bInsertA or self.bReplace:
                 bFind = True
             # loop through all items in the file list
-            for item in oldfilelist:
-                # get index position of item in oldfilelist to crossreference with other file lists
-                indexposn = oldfilelist.index(item)
+            for item in self.oldFileList:
+                # get index position of item in oldFileList to crossreference with other file lists
+                indexposn = self.oldFileList.index(item)
                 basefilename = os.path.splitext(item)[0]
                 basefileext = os.path.splitext(item)[1]
                 # if required to find text for replacement or insertion, do so
                 tempfilename = basefilename
-                tempfilepathlist[indexposn] = subdirlist[indexposn] + tempfilename + basefileext
+                self.tempFilePathList[indexposn] = self.subDirList[indexposn] + tempfilename + basefileext
                 if bFind:
                     textposn = item.find(texttofind)
                     # if the provided text was found, do replacement or insertion as necessary
                     if textposn > -1:
                         # if user chose to insert after text, do so
-                        if Example.bInsertA:
+                        if self.bInsertA:
                             textposn = tempfilename.find(texttofind)
                             textleft = tempfilename[0:textposn+len(texttofind)]
                             textposnend = textposn + len(texttofind)
                             textright = tempfilename[textposnend:len(tempfilename)]
                             tempfilename = textleft + sInsertA + textright
-                            newfilelist[indexposn] = tempfilename
-                            newfilepathlist[indexposn] = subdirlist[indexposn] + tempfilename + basefileext
+                            self.newFileList[indexposn] = tempfilename
+                            self.newFilePathList[indexposn] = self.subDirList[indexposn] + tempfilename + basefileext
                             saveFile(indexposn)
-                            tempfilepathlist[indexposn] = subdirlist[indexposn] + tempfilename + basefileext
+                            self.tempFilePathList[indexposn] = self.subDirList[indexposn] + tempfilename + basefileext
 
                         # if user chose to insert before text, do so
-                        if Example.bInsertB:
+                        if self.bInsertB:
                             textposn = tempfilename.find(texttofind)
                             textleft = tempfilename[0:textposn]
                             textposnend = textposn
                             textright = tempfilename[textposnend:len(tempfilename)]
                             tempfilename = textleft + sInsertB + textright
-                            newfilelist[indexposn] = tempfilename
-                            newfilepathlist[indexposn] = subdirlist[indexposn] + tempfilename + basefileext
+                            self.newFileList[indexposn] = tempfilename
+                            self.newFilePathList[indexposn] = self.subDirList[indexposn] + tempfilename + basefileext
                             saveFile(indexposn)
-                            tempfilepathlist[indexposn] = subdirlist[indexposn] + tempfilename + basefileext
+                            self.tempFilePathList[indexposn] = self.subDirList[indexposn] + tempfilename + basefileext
 
                         # if user chose to replace text, do so
-                        if Example.bReplace:
+                        if self.bReplace:
                             textposn = tempfilename.find(texttofind)
                             textleft = tempfilename[0:textposn]
                             textposnend = textposn + len(texttofind)
                             textright = tempfilename[textposnend:len(tempfilename)]
                             tempfilename = textleft + sReplace + textright
-                            newfilelist[indexposn] = tempfilename
-                            newfilepathlist[indexposn] = subdirlist[indexposn] + tempfilename + basefileext
+                            self.newFileList[indexposn] = tempfilename
+                            self.newFilePathList[indexposn] = self.subDirList[indexposn] + tempfilename + basefileext
                             saveFile(indexposn)
-                            tempfilepathlist[indexposn] = subdirlist[indexposn] + tempfilename + basefileext
+                            self.tempFilePathList[indexposn] = self.subDirList[indexposn] + tempfilename + basefileext
 
                 # if user chose to add suffix, do so
-                if Example.bSuffix:
+                if self.bSuffix:
                     tempfilename = tempfilename + sSuffix
-                    newfilelist[indexposn] = tempfilename
-                    newfilepathlist[indexposn] = subdirlist[indexposn] + tempfilename + basefileext
+                    self.newFileList[indexposn] = tempfilename
+                    self.newFilePathList[indexposn] = self.subDirList[indexposn] + tempfilename + basefileext
                     saveFile(indexposn)
-                    tempfilepathlist[indexposn] = subdirlist[indexposn] + tempfilename + basefileext
+                    self.tempFilePathList[indexposn] = self.subDirList[indexposn] + tempfilename + basefileext
 
                 # if user chose to add prefix, do so
-                if Example.bPrefix:
+                if self.bPrefix:
                     tempfilename = sPrefix + tempfilename
-                    newfilelist[indexposn] = tempfilename
-                    newfilepathlist[indexposn] = subdirlist[indexposn] + tempfilename + basefileext
+                    self.newFileList[indexposn] = tempfilename
+                    self.newFilePathList[indexposn] = self.subDirList[indexposn] + tempfilename + basefileext
                     saveFile(indexposn)
-                    tempfilepathlist[indexposn] = subdirlist[indexposn] + tempfilename + basefileext
+                    self.tempFilePathList[indexposn] = self.subDirList[indexposn] + tempfilename + basefileext
 
-                snewfilelist.set("\n".join(str(x) for x in newfilelist))
-
-
-        # function for getting all files within a directories and its subs
-        def getfilelist(directory):
-            subdirlist.clear()
-            oldfilelist.clear()
-            oldfilepathlist.clear()
-            newfilelist.clear()
-            newfilepathlist.clear()
-            # clear new file name display
-            snewfilelist.set("")
-            for subdir, dirs, files in os.walk(directory):
-                for file in files:
-                    filepath = subdir + os.sep + file
-                    # add subdir + ending slash to global subdir variable
-                    subdirlist.append(subdir + os.sep)
-                    # add file to global list variable
-                    oldfilelist.append(file)
-                    # append null string to keep list the same size
-                    newfilelist.append("")
-                    oldfilepathlist.append(filepath)
-                    newfilepathlist.append("")
-                    tempfilepathlist.append("")
-                if not Example.bIncludeSubdirs:
-                    # if user selected not to include subdirectories,
-                    #  break out of loop after going through files
-                    break
-            # get string list of files to display to the user
-            soldfilelist.set("\n".join(str(x) for x in oldfilelist))
-            # get string list of filepaths to display to the user
-            soldfilepathlist.set("\n".join(str(x) for x in oldfilepathlist))
-            updatebody(scrollable_body)
+                snewFileList.set("\n".join(str(x) for x in self.newFileList))
 
 
-        # function for browsing to a directory
-        def BrowseFolder():
-            # prompt user
-            global userdirectory
-            userdirectory=filedialog.askdirectory()
-            # check if user cancelled prompt - this doesn't actually work (always returns folderPath <> None)
-            if not folderPath is None:
-                folderPath.set(userdirectory)
-                getfilelist(userdirectory)
-            else:
-                folderPath.set("nothing selected")
 
 
-        def updatebody(bdy):
-            bdy.update()
-
+        # functions that will get used to rename files
+        def saveFile(i):
+            os.rename(self.tempFilePathList[i], self.newFilePathList[i])
 
         # function to toggle the relief state of a passed button
         def togglebuttonrelief(btn):
@@ -189,45 +188,50 @@ class Example(Frame):
 
         # function to toggle bIncludeSubdirs global boolean and button display state
         def togglebIncludeSubdirs(btn):
-            global userdirectory
+            global userDirectory
             togglebuttonrelief(btn)
-            Example.bIncludeSubdirs = not Example.bIncludeSubdirs
-            getfilelist(userdirectory)
+            self.bIncludeSubdirs = not self.bIncludeSubdirs
+            getfilelist(userDirectory)
 
         # function to toggle bPrefix global boolean and button display state
         def togglebPrefix(btn):
             togglebuttonrelief(btn)
-            Example.bPrefix = not Example.bPrefix
+            self.bPrefix = not self.bPrefix
 
         # function to toggle bSuffix global boolean and button display state
         def togglebSuffix(btn):
             togglebuttonrelief(btn)
-            Example.bSuffix = not Example.bSuffix
+            self.bSuffix = not self.bSuffix
 
         # function to toggle bInsertB global boolean and button display state
         def togglebInsertB(btn):
             togglebuttonrelief(btn)
-            Example.bInsertB = not Example.bInsertB
+            self.bInsertB = not self.bInsertB
 
         # function to toggle bInsertA global boolean and button display state
         def togglebInsertA(btn):
             togglebuttonrelief(btn)
-            Example.bInsertA = not Example.bInsertA
+            self.bInsertA = not self.bInsertA
 
         # function to toggle bReplace global boolean and button display state
         def togglebReplace(btn):
             togglebuttonrelief(btn)
-            Example.bReplace = not Example.bReplace
+            self.bReplace = not self.bReplace
 
         def runRenameStuff():
-            prefix = entryPrefix.get()
-            suffix = entrySuffix.get()
-            insertB = entryInsertB.get()
-            insertA = entryInsertA.get()
-            replace = entryReplace.get()
-            findtext = entryFindText.get()
-            RenameStuff(prefix, suffix, insertB, insertA, replace, findtext)
+            self.prefix = entryPrefix.get()
+            self.suffix = entrySuffix.get()
+            self.insertB = entryInsertB.get()
+            self.insertA = entryInsertA.get()
+            self.replace = entryReplace.get()
+            self.findtext = entryFindText.get()
+            RenameStuff(self.prefix, self.suffix, self.insertB, self.insertA, self.replace, self.findtext)
 
+        def updatebody(bdy):
+            bdy.update()
+
+
+        # actual GUI
         self.master.title("Directory Info")
         # main window
         self.pack(fill=BOTH, expand=True,padx=5,pady=5)
@@ -250,7 +254,7 @@ class Example(Frame):
         label2 = Label(frame2, text="Directory Selected:", width=20)
         label2.pack(side=LEFT, padx=5, pady=5)
         # list user selected directory to the right
-        label2a = Label(frame2, textvariable=folderPath)
+        label2a = Label(frame2, textvariable=sUserDirectory)    #this is not working
         label2a.pack(fill=X, padx=5, expand=True)
 
         # fourth frame to show options for renaming
@@ -266,43 +270,47 @@ class Example(Frame):
         buttonStartRenaming = Button(frame4a, text="Rename some shit!", width=20,\
                 command=lambda: runRenameStuff())
         buttonStartRenaming.pack(side=RIGHT)
+        # create buttons for user to toggle rename functions and create entry boxes for user to provide matching text strings
+        # Prefix
         frame4b = Frame(frame4, borderwidth=2, relief="solid")
         frame4b.pack(fill=X)
-        frame4c = Frame(frame4, borderwidth=2, relief="solid")
-        frame4c.pack(fill=X)
-        frame4d = Frame(frame4, borderwidth=2, relief="solid")
-        frame4d.pack(fill=X)
-        frame4e = Frame(frame4, borderwidth=2, relief="solid")
-        frame4e.pack(fill=X)
-        frame4f = Frame(frame4, borderwidth=2, relief="solid")
-        frame4f.pack(fill=X)
-        frame4g = Frame(frame4, borderwidth=2, relief="solid")
-        frame4g.pack(fill=X)
-        # create buttons for user to toggle rename functions
         buttonPrefix = tk.Button(frame4b, text="Add Prefix", width=12, relief="raised", command=lambda: togglebPrefix(buttonPrefix))
         buttonPrefix.pack(side=LEFT, padx=5)
-        buttonSuffix = tk.Button(frame4c, text="Add Suffix", width=12, relief="raised", command=lambda: togglebSuffix(buttonSuffix))
-        buttonSuffix.pack(side=LEFT, padx=5)
-        buttonInsertB = tk.Button(frame4d, text="Insert Before", width=12, relief="raised", command=lambda: togglebInsertB(buttonInsertB))
-        buttonInsertB.pack(side=LEFT, padx=5)
-        buttonInsertA = tk.Button(frame4e, text="Insert After", width=12, relief="raised", command=lambda: togglebInsertA(buttonInsertA))
-        buttonInsertA.pack(side=LEFT, padx=5)
-        buttonReplace = tk.Button(frame4f, text="Replace", width=12, relief="raised", command=lambda: togglebReplace(buttonReplace))
-        buttonReplace.pack(side=LEFT, padx=5)
-        # create label for text to find
-        labelFind = Label(frame4g, text="Text to find (for Insert Before, Insert After, and Replace):")
-        labelFind.pack(side=LEFT,padx=5)
-        # create entry boxes for user to provide text strings
         entryPrefix = Entry(frame4b)
         entryPrefix.pack(padx=5)
+        # Suffix
+        frame4c = Frame(frame4, borderwidth=2, relief="solid")
+        frame4c.pack(fill=X)
+        buttonSuffix = tk.Button(frame4c, text="Add Suffix", width=12, relief="raised", command=lambda: togglebSuffix(buttonSuffix))
+        buttonSuffix.pack(side=LEFT, padx=5)
         entrySuffix = Entry(frame4c)
         entrySuffix.pack(padx=5)
+        # Insert Before
+        frame4d = Frame(frame4, borderwidth=2, relief="solid")
+        frame4d.pack(fill=X)
+        buttonInsertB = tk.Button(frame4d, text="Insert Before", width=12, relief="raised", command=lambda: togglebInsertB(buttonInsertB))
+        buttonInsertB.pack(side=LEFT, padx=5)
         entryInsertB = Entry(frame4d)
         entryInsertB.pack(padx=5)
+        # Insert After
+        frame4e = Frame(frame4, borderwidth=2, relief="solid")
+        frame4e.pack(fill=X)
+        buttonInsertA = tk.Button(frame4e, text="Insert After", width=12, relief="raised", command=lambda: togglebInsertA(buttonInsertA))
+        buttonInsertA.pack(side=LEFT, padx=5)
         entryInsertA = Entry(frame4e)
         entryInsertA.pack(padx=5)
+        # Replace
+        frame4f = Frame(frame4, borderwidth=2, relief="solid")
+        frame4f.pack(fill=X)
+        buttonReplace = tk.Button(frame4f, text="Replace", width=12, relief="raised", command=lambda: togglebReplace(buttonReplace))
+        buttonReplace.pack(side=LEFT, padx=5)
         entryReplace = Entry(frame4f)
         entryReplace.pack(padx=5)
+        # Text to Find
+        frame4g = Frame(frame4, borderwidth=2, relief="solid")
+        frame4g.pack(fill=X)
+        labelFind = Label(frame4g, text="Text to find (for Insert Before, Insert After, and Replace):")
+        labelFind.pack(side=LEFT,padx=5)
         entryFindText = Entry(frame4g)
         entryFindText.pack(padx=5)
 
@@ -315,23 +323,22 @@ class Example(Frame):
         frame3a.pack(side=LEFT, fill=BOTH, expand=True)
         label3a1 = Label(frame3a, text="Old File Name List", width=20, anchor="center", borderwidth=2, relief="solid")
         label3a1.pack(anchor="center", padx=5, pady=5)
-        label3a2 = Label(frame3a, textvariable=soldfilelist)
+        label3a2 = Label(frame3a, textvariable=soldFileList)
         label3a2.pack(padx=2, pady=5)
         # sub frame in the middle to show new filenames
         frame3b = Frame(scrollable_body, borderwidth=2, relief="raised")
         frame3b.pack(side=LEFT, fill=BOTH, expand=True)
         label3b1 = Label(frame3b, text="New File Name", width=20, anchor="center", borderwidth=2, relief="solid")
         label3b1.pack(anchor="center", padx=5, pady=5)
-        label3b2 = Label(frame3b, textvariable=snewfilelist)
+        label3b2 = Label(frame3b, textvariable=snewFileList)
         label3b2.pack(padx=2, pady=5)
         # sub frame on the right to show filenames with paths
         frame3c = Frame(scrollable_body, borderwidth=2, relief="solid")
         frame3c.pack(side=LEFT, fill=BOTH, expand=True)
         label3c1 = Label(frame3c, text="File Path List", width=15, anchor="center", borderwidth=2, relief="solid")
         label3c1.pack(anchor="center", padx=5, pady=5)
-        label3c2 = Label(frame3c, textvariable=soldfilepathlist)
+        label3c2 = Label(frame3c, textvariable=soldFilePathList)
         label3c2.pack(padx=2, pady=5)
-
 
         scrollable_body.update()
 
@@ -376,63 +383,38 @@ class Scrollable(ttk.Frame):
 
 
 def main():
-
     # main window
     root = Tk()
     # set size of window
     root.geometry("1000x800+300+200")
 
-    # initialize a bunch of global variables
-
-    # declare global directory to make it available in the rename function
-    global userdirectory
-    userdirectory= ""
-    # declare global variable to make the data available
-    global folderPath
+    #initialize any global variables
+    global userDirectory
     # this apparently needs to be here so that labels can access the data
-    folderPath = tk.StringVar()
-    # global variable to hold list of files found
-    global subdirlist
-    # initialize as list type
-    subdirlist = []
-    # global variable to hold list of files found
-    global oldfilelist
-    # initialize as list type
-    oldfilelist = []
+    userDirectory = tk.StringVar()
+    #initialize any global variables
+    global sUserDirectory
+    # this apparently needs to be here so that labels can access the data
+    sUserDirectory = tk.StringVar()
     # global variable to hold string list of files found
-    global soldfilelist
+    global soldFileList
     # this apparently needs to be here so that labels can access the data
-    soldfilelist = tk.StringVar()
-    # global variable to hold list of files found
-    global newfilelist
-    # initialize as list type
-    newfilelist = []
+    soldFileList = tk.StringVar()
     # global variable to hold string list of files found
-    global snewfilelist
+    global snewFileList
     # this apparently needs to be here so that labels can access the data
-    snewfilelist = tk.StringVar()
-    # global variable to hold list of filepaths found
-    global oldfilepathlist
-    # initialize as list type
-    oldfilepathlist = []
+    snewFileList = tk.StringVar()
     # global variable to hold string list of files found
-    global soldfilepathlist
+    global soldFilePathList
     # this apparently needs to be here so that labels can access the data
-    soldfilepathlist = tk.StringVar()
-    global newfilepathlist
-    # initialize as list type
-    newfilepathlist = []
+    soldFilePathList = tk.StringVar()
     # global variable to hold string list of files found
-    global snewfilepathlist
+    global snewFilePathList
     # this apparently needs to be here so that labels can access the data
-    snewfilepathlist = tk.StringVar()
-    # global variable to hold temporary list of file paths as files are renamed
-    global tempfilepathlist
-    tempfilepathlist = []
-
+    snewFilePathList = tk.StringVar()
 
     # create frame in root
-    app = Example()
+    app = FileRename()
     # a button to close root
     button = tk.Button(root, text="Close", width=15,\
             command=root.destroy, borderwidth=5, relief="solid")
@@ -440,6 +422,10 @@ def main():
     button.pack(pady=15)
     # end main
     root.mainloop()
+
+
+
+
 
 # just here for reasons, to get things started
 if __name__ == '__main__':
